% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pactr-class_cloning.R
\name{clone}
\alias{clone}
\title{Clones the mpactr object}
\usage{
clone(mpactr_object)
}
\arguments{
\item{mpactr_object}{the object that we are cloning.}
}
\description{
This function allows us to clone the mpactr object to give our users a more r-based flow.
}
