% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cultures_data}
\alias{cultures_data}
\title{LC-MS/MS sample data}
\format{
\subsection{\code{culture_data}}{

A \code{mpactr} with 2 attributes:
\describe{
\item{peak_table}{A feature table of class \code{data.table}}
\item{meta_data}{A \code{data.table} with associated sample metadata}
}
}
}
\usage{
cultures_data
}
\value{
An \code{mpactr} R6 class object.
}
\description{
A \code{mpactr} R6 class object containing a feature table and associated sample
metadata.
}
\keyword{datasets}
