% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{example}
\alias{example}
\title{Get file paths for examples}
\usage{
example(file = NULL)
}
\arguments{
\item{file}{Name of a file. If \code{NULL}, all examples files will be listed.}
}
\description{
mpactr contains a number of example files in the \verb{inst\\extdata} directory. This function makes them accessible in
that shows how file paths are used in function examples.
}
\examples{
example()

example("metadata.csv")
}
