% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{filter_mismatch_ions}
\alias{filter_mismatch_ions}
\title{Mispicked ions filter}
\usage{
filter_mismatch_ions(
  mpactr_object,
  ringwin = 0.5,
  isowin = 0.01,
  trwin = 0.005,
  max_iso_shift = 3,
  merge_peaks = TRUE
)
}
\arguments{
\item{isowin}{Isotopic mass window. Defualt = 0.01.}

\item{trwin}{A \code{numeric} denoting the retention time threhold for assessing if ions should be merged. Defulat = 0.005.}

\item{max_iso_shift}{A \code{numeric}. Default = 3.}

\item{merge_peaks}{A \code{logical} to determine if peaks found to belong to the same ion should be merged in the feature table.}

\item{mpact_object}{An \code{mpactr_object}. See \code{\link[=import_data]{import_data()}}.}

\item{ringwing}{Ringing mass window. Default = 0.5.}
}
\value{
an \code{mpactr_object}
}
\description{
\code{filter_mismatch_ions()} identifies ions that were incorrectly split into separate features during preprocessing. This filter checks the feature table for similar ions in terms of mass and retention time. Peaks found to be similar are merged into a single feature given \code{merge_peaks} is \code{TRUE}.
}
\examples{
data <- import_data("extdata/coculture_peak_table.csv", 
                    "extdata/metadata.csv")

data_filter <- filter_mismatch_ions(data,
                              ringwin = 0.5, 
                              isowin = 0.01,
                              trwin = 0.005,
                              max_iso_shift = 3,
                              remove_ions = TRUE)

}
