% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpactr-accessors.R
\name{get_peak_table}
\alias{get_peak_table}
\title{Return the peak table data.table from the mpactr object.}
\usage{
get_peak_table(mpactr_object)
}
\arguments{
\item{mpactr_object}{The mpactr object that is created by calling
the import_data() function.}
}
\value{
a \code{data.table}.
}
\description{
\code{get_peak_table()} a wrapper function to return the peak table
object of the given mpactr object.
}
\examples{
data <- import_data(
  example_path("coculture_peak_table.csv"),
  example_path("metadata.csv"),
  format = "Progenesis"
)

peak_table <- get_peak_table(data)

}
