% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_pactr-accessors.R
\name{group_averages}
\alias{group_averages}
\title{Get groups averages.}
\usage{
group_averages(mpactr_object)
}
\arguments{
\item{mpactr_object}{The mpactr object that is created by calling the import_data() function.}
}
\value{
a \code{data.table} retporting the average and relative standard deviation across biological groups and techincal replicates within each group].
}
\description{
\code{group_averages()} is a wrapper function to return group averages for calculated with \code{\link[=filter_group]{filter_group()}}.
}
\examples{
data <- import_data(example("coculture_peak_table.csv"),
                    example("metadata.csv"))

data_filter <- filter_group(data, group_to_remove = "Blanks")

group_averages <- group_averages(data_filter)
head(group_averages)

}
