% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{limit_cores}
\alias{limit_cores}
\title{Limit Cores}
\usage{
limit_cores()
}
\value{
does not return anything.
}
\description{
Limits the amount of cores used to two. Mostly used in test
and examples so we can pass cran checks.
}
\examples{
limit_cores()

}
