% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpactR-package.R
\docType{package}
\name{mpactR-package}
\alias{mpactR}
\alias{mpactR-package}
\title{mpactR: Filters for the Correction of Preprocessed MS Data}
\description{
A fast tool for the correction of pre-processed tandem MS/MS features.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://www.schlosslab.org/mpactR/}
}

}
\author{
\strong{Maintainer}: Patrick Schloss \email{pschloss@umich.edu} (\href{https://orcid.org/0000-0002-6935-4275}{ORCID})

Authors:
\itemize{
  \item Allison Mason \email{masonar@umich.edu} (\href{https://orcid.org/0000-0003-1339-1592}{ORCID})
  \item Gregory Johnson \email{grejoh@umich.edu}
}

}
\keyword{internal}
