% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{filter_cv}
\alias{filter_cv}
\title{Filter Non-reproducible ions}
\usage{
filter_cv(mpactr_object, cv_threshold = NULL, cv_param, copy_object = FALSE)
}
\arguments{
\item{mpactr_object}{An \code{mpactr_object}. See \code{\link[=import_data]{import_data()}}.}

\item{cv_threshold}{Coefficient of variation threshold.}

\item{cv_param}{Coefficient of variation (CV) to use for filtering. Options are "mean" or "median", corresponding to mean and median CV, respectively.}

\item{copy_object}{A \code{boolean} paramter that allows users to return a copied object instead of modifying the object.}
}
\value{
an \code{mpactr_object}
}
\description{
\code{filter_cv()} removes feature ions that are found to be non-reproducible between technical inejction replicates. Replicability is assessed via mean or median coefficient of variation (CV) between technical replicates. As such, this fitler is expecting an input dataset with at least two replicate injections per sample.
}
\examples{

data <- import_data(example("coculture_peak_table.csv"),
                    example("metadata.csv"))

data_filter <- filter_cv(data,
                              cv_threshold = 0.01,
                              cv_param = "mean")

data_filter <- filter_cv(data,
                              cv_threshold = 0.01,
                              cv_param = "median")

}
