% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{filter_group}
\alias{filter_group}
\title{Filter Ions by Group}
\usage{
filter_group(
  mpactr_object,
  group_threshold = 0.01,
  group_to_remove,
  remove_ions = TRUE
)
}
\arguments{
\item{mpactr_object}{An \code{mpactr_object}. See \code{\link[=import_data]{import_data()}}.}

\item{group_threshold}{Relative abundance threshold at which to remove ions. Default = 0.01.}

\item{group_to_remove}{Biological group name to remove ions from.}

\item{remove_ions}{A \code{logical}. If \code{TRUE} failing ions will be removed from the peak table. Default = TRUE.}
}
\value{
an \code{mpactr_object}
}
\description{
Filter Ions by Group
}
\details{
\code{filter_group()} removes ions which have a signigifanct presence in the defined group (typically solvent blanks). Significance is determined by relative ion abundance arcoss groups at the user-defined \code{group_threshold}.

Filtering is assessed on function parameter \code{group_threshold}. The defualt is 0.01, meaning features present in \code{group_to_remove} at greater than 1\% are removed.
}
\examples{
data <- import_data(example("coculture_peak_table.csv"),
                    example("metadata.csv"))

data_filter <- filter_group(data,
                              group_threshold = 0.01,
                              group_to_remove = "Blanks",
                              remove_ions = TRUE)

}
