% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{filter_insource_ions}
\alias{filter_insource_ions}
\title{Filter Insource ions}
\usage{
filter_insource_ions(mpactr_object, cluster_threshold = 0.95)
}
\arguments{
\item{mpactr_object}{An \code{mpactr_object}. See \code{\link[=import_data]{import_data()}}.}

\item{cluster_threshold}{cluster threshold for ion deconvolution. Default = 0.95.}
}
\value{
an \code{mpactr_object}
}
\description{
\code{filter_insource_ions()} determines insource ion fragments deconvolution via hierarchical clustering. Highly correlated ions with the same retention times are identified and removed.
}
\examples{
data <- import_data(load(system.file('extdata', 'coculture_peak_table.csv', package = 'mpactr')), 
                      load(system.file('extdata', 'metadata.csv', package = 'mpactr')))

data_filter <- filter_insource_ions(data,
                              cluster_threhold = 0.95)

}
