% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{import_data}
\alias{import_data}
\title{Import data into an mpactr object.}
\usage{
import_data(peak_table_file_path, meta_data_file_path)
}
\arguments{
\item{peak_table_file_path}{The file path to your feature table file.}

\item{meta_data_file_path}{The file path to your meta_data file.}
}
\value{
an \code{mpactr_object}
}
\description{
\code{import_data()} takes two file paths, one for the pre-processed feature table and one for sample metadata. Both files should be .csv.
}
\details{
Expected feature table format.

Expected metadata table format.
}
\examples{
data <- import_data(example("coculture_peak_table.csv"),
                    example("metadata.csv"))

}
