% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{import_data}
\alias{import_data}
\title{Import data into an mpactr object.}
\usage{
import_data(peak_table, meta_data, format = "none")
}
\arguments{
\item{peak_table}{The file path to your feature table file.}

\item{meta_data}{The file path to your meta_data file or \code{data.frame}.}

\item{format}{The expected exported type of your peak table, can be one of Progenesis, MzMine, Metaboscape, or None.}
}
\value{
an \code{mpactr_object}
}
\description{
\code{import_data()} takes two file paths, one for the pre-processed feature table and one for sample metadata. Both files should be .csv.
}
\details{
mpactR requires 2 files as imput: a feature table and metadata file. Both are expected to be comma separated files (\emph{.csv}).
\enumerate{
\item peak_table: a peak table in Progenesis format is expected. To export a compatable peak table in Progenesis, navigate to the
\emph{Review Compounds} tab then File -> Export Compound Measurements. Select the following properties: Compound, m/z,
Retention time (min), and Raw abundance and click ok.
\item metadata: a table with sample information. At minimum the following columns are expected: Injection, Sample_Code,
Biological_Group. Injection is the sample name and is expected to match sample column names in the peak_table. Sample_Code
is the id for technical replicate groups. Biological Group is the id for biological replicate groups. Other sample metadata
can be added, and is encouraged for downstream analysis following filtering with mpactR.
}
}
\examples{
data <- import_data(example("coculture_peak_table.csv"),
                    example("metadata.csv"),
                    format = "Progenesis")

meta_data <- read.csv(example("metadata.csv"))
data <- import_data(example("coculture_peak_table.csv"),
                    meta_data,
                    format = "Progenesis")

}
