% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpactR-package.R
\docType{package}
\name{mpactr-package}
\alias{mpactr}
\alias{mpactr-package}
\title{mpactr: Correction of Preprocessed MS Data}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

An 'R' implementation of the 'python' program Metabolomics Peak Analysis Computational Tool ('MPACT') (Robert M. Samples, Sara P. Puckett, and Marcy J. Balunas (2023) \doi{10.1021/acs.analchem.2c04632}). Filters in the package serve to address common errors in tandem mass spectrometry preprocessing, including: (1) isotopic patterns that are incorrectly split during preprocessing, (2) features present in solvent blanks due to carryover between samples, (3) features whose abundance is greater than user-defined abundance threshold in a specific group of samples, for example media blanks, (4) ions that are inconsistent between technical replicates, and (5) in-source fragment ions created during ionization before fragmentation in the tandem mass spectrometry workflow.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://www.mums2.org/mpactr/}
  \item \url{https://github.com/mums2/mpactr}
  \item Report bugs at \url{https://github.com/mums2/mpactr/issues}
}

}
\author{
\strong{Maintainer}: Patrick Schloss \email{pschloss@umich.edu} (\href{https://orcid.org/0000-0002-6935-4275}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Allison Mason \email{masonar@umich.edu} (\href{https://orcid.org/0000-0003-1339-1592}{ORCID})
  \item Gregory Johnson \email{grejoh@umich.edu} (\href{https://orcid.org/0009-0008-3890-0297}{ORCID})
}

}
\keyword{internal}
